+++ packages/errors/tsdown.config.ts
@@ -1,3 +1,12 @@
-import baseConfig from '../tsdown.config.ts';
+import { defineConfig, type UserConfig } from 'tsdown';
 
-export default baseConfig;
+import baseConfig from '../../tsdown.config.json' with { type: 'json' };
+
+const config: UserConfig = defineConfig({
+  ...(baseConfig as UserConfig),
+  entry: {
+    index: 'src/index.ts',
+  },
+});
+
+export default config;
