+++ plugins/logrotator/tsdown.config.ts
@@ -1,3 +1,10 @@
-import pluginConfig from '../tsdown.config.ts';
+import { defineConfig, type UserConfig } from 'tsdown';
 
-export default pluginConfig;
+import baseConfig from '../../tsdown.config.json' with { type: 'json' };
+
+const config: UserConfig = defineConfig({
+  ...(baseConfig as UserConfig),
+  entry: 'src/**/*.ts',
+});
+
+export default config;
