+++ plugins/view/test/fixtures/apps/ts/app.ts
@@ -1,12 +1,14 @@
 import { Application } from 'egg';
 
+import { ViewEngineBase } from '../../../../src/index.ts';
+
 const setupApp = (app: Application): void => {
   app.view.use('newEngine', NewEngine);
 };
 
 export default setupApp;
 
-class NewEngine {
+class NewEngine extends ViewEngineBase {
   render(): Promise<string> {
     return Promise.resolve('');
   }
