+++ pnpm-workspace.yaml
@@ -20,6 +20,7 @@ catalog:
   '@types/accepts': ^1.3.7
   '@types/body-parser': ^1.19.5
   '@types/bytes': ^3.1.5
+  '@types/common-tags': ^1.8.4
   '@types/content-disposition': ^0.5.8
   '@types/content-type': ^1.1.8
   '@types/cookie-parser': ^1.4.8
@@ -68,6 +69,7 @@ catalog:
   bytes: ^3.1.2
   c8: ^10.1.3
   cache-content-type: ^2.0.0
+  cheerio: ^1.0.0
   camelcase: ^8.0.0
   cfork: ^2.0.0
   ci-parallel-vars: ^1.0.1
